TAMAYA-324: Fixes for test failures provided by William Lieurance
First, the CDIAwareServiceContextTest now uses the same Arquillian
backing as various other successful tests.

Second, the ConfiguredVetoExtensionTest is now more idempotent with respect
to the ConfigurationProvider.  Previously, it would leave the shared
ConfigurationProvider pointing at the mocked ConfigurationContext, which
caused other tests (specifically ConfiguredBTest) which rely on a valid
ConfigurationContext to fail.